@@ -1,1 +1,1 @@
-2.6.6
+3.0.6
