@@ -6,7 +6,9 @@ class Manage::AgreementsController < Manage::ApplicationController
 
   # GET /agreements
   def index
-    @agreements = Agreement.all
+    @agreements_search = Agreement.ransack(params[:agreements_search], search_key: :agreements_search)
+    @agreements = @agreements_search.result(distinct: true)
+    @agreements_pagy, @agreements = pagy(@agreements, page_param: 'agreements_page', items: 10)
   end
 
   # GET /agreements/new
@@ -28,7 +30,7 @@ class Manage::AgreementsController < Manage::ApplicationController
 
   # PATCH/PUT /agreements/1
   def update
-    @agreement.update_attributes(agreement_params)
+    @agreement.update(agreement_params)
     flash[:notice] = nil
     redirect_to manage_agreements_path
   end
