@@ -7,11 +7,9 @@ class Manage::MessagesController < Manage::ApplicationController
   respond_to :html, :json
 
   def index
-    respond_with(:manage, Message.all)
-  end
-
-  def datatable
-    render json: BulkMessageDatatable.new(params, view_context: view_context)
+    @bulk_messages_search = Message.ransack(params[:bulk_messages_search], search_key: :bulk_messages_search)
+    @bulk_messages = @bulk_messages_search.result(distinct: true)
+    @bulk_messages_pagy, @bulk_messages = pagy(@bulk_messages, page_param: 'bulk_messages_page', items: 10)
   end
 
   def show
@@ -35,7 +33,7 @@ class Manage::MessagesController < Manage::ApplicationController
   end
 
   def update
-    @message.update_attributes(message_params)
+    @message.update(message_params)
     respond_with(:manage, @message)
   end
 
@@ -73,7 +71,7 @@ class Manage::MessagesController < Manage::ApplicationController
 
   def duplicate
     new_message = @message.dup
-    new_message.update_attributes(
+    new_message.update(
       delivered_at: nil,
       started_at: nil,
       queued_at: nil,
@@ -96,7 +94,7 @@ class Manage::MessagesController < Manage::ApplicationController
   def template_update
     message_template = MessageTemplate.uncached_instance
     message_template_params = params.require(:message_template).permit(:html)
-    message_template.update_attributes(message_template_params)
+    message_template.update(message_template_params)
     redirect_to template_manage_messages_path
   end
 
