@@ -14,11 +14,6 @@ class Manage::MessagesControllerTest < ActionController::TestCase
       assert_redirected_to new_user_session_path
     end
 
-    should "not allow access to manage_messages datatables api" do
-      post :datatable, format: :json, params: { "columns[0][data]" => "" }
-      assert_response 401
-    end
-
     should "not allow access to manage_messages#new" do
       get :new
       assert_response :redirect
@@ -109,12 +104,6 @@ class Manage::MessagesControllerTest < ActionController::TestCase
 
     should "not allow access to manage_messages#index" do
       get :index
-      assert_response :redirect
-      assert_redirected_to root_path
-    end
-
-    should "not allow access to manage_messages datatables api" do
-      post :datatable, format: :json, params: { "columns[0][data]" => "" }
       assert_response :redirect
       assert_redirected_to root_path
     end
@@ -210,13 +199,7 @@ class Manage::MessagesControllerTest < ActionController::TestCase
     should "not allow access to manage_messages#index" do
       get :index
       assert_response :redirect
-      assert_redirected_to manage_checkins_path
-    end
-
-    should "not allow access to manage_messages datatables api" do
-      post :datatable, format: :json, params: { "columns[0][data]" => "" }
-      assert_response :redirect
-      assert_redirected_to manage_checkins_path
+      assert_redirected_to manage_root_path
     end
 
     should "not allow access to manage_messages#new" do
@@ -228,7 +211,7 @@ class Manage::MessagesControllerTest < ActionController::TestCase
     should "not allow access to manage_messages#show" do
       get :show, params: { id: @message }
       assert_response :redirect
-      assert_redirected_to manage_checkins_path
+      assert_redirected_to manage_root_path
     end
 
     should "not allow access to manage_messages#edit" do
@@ -266,7 +249,7 @@ class Manage::MessagesControllerTest < ActionController::TestCase
     should "not allow access to manage_messages#preview" do
       get :preview, params: { id: @message }
       assert_response :redirect
-      assert_redirected_to manage_checkins_path
+      assert_redirected_to manage_root_path
     end
 
     should "not allow access to manage_messages#live_preview" do
@@ -309,11 +292,6 @@ class Manage::MessagesControllerTest < ActionController::TestCase
 
     should "allow access to manage_messages#index" do
       get :index
-      assert_response :success
-    end
-
-    should "allow access to manage_messages datatables api" do
-      post :datatable, format: :json, params: { "columns[0][data]" => "" }
       assert_response :success
     end
 
@@ -531,7 +509,7 @@ class Manage::MessagesControllerTest < ActionController::TestCase
       end
 
       should "reset status" do
-        @message.update_attributes(
+        @message.update(
           delivered_at: Time.now,
           started_at: Time.now,
           queued_at: Time.now,
@@ -541,7 +519,7 @@ class Manage::MessagesControllerTest < ActionController::TestCase
       end
 
       should "maintain similar fields" do
-        @message.update_attributes(
+        @message.update(
           name: "My message name",
           subject: "This subject",
           body: "hello world",
