@@ -9,13 +9,13 @@ class Manage::ApplicationController < ApplicationController
   end
 
   def require_director
-    return redirect_to manage_checkins_path if current_user.volunteer?
+    return redirect_to manage_root_path if current_user.volunteer?
     return redirect_to manage_root_path if current_user.organizer?
     return redirect_to root_path unless current_user.try(:director?)
   end
 
   def require_director_or_organizer
-    return redirect_to manage_checkins_path if current_user.volunteer?
+    return redirect_to manage_root_path if current_user.volunteer?
     return redirect_to root_path unless current_user.organizing_staff?
   end
 
