@@ -1,13 +1,19 @@
 class School < ApplicationRecord
+  include PgSearch::Model
+  pg_search_scope :search_by_name, against: :name, using: { tsearch: { prefix: true } }
   audited
 
   validates_presence_of :name
 
-  validates_uniqueness_of :name
+  validates_uniqueness_of :name, case_sensitive: false
 
   strip_attributes
 
   has_many :questionnaires
+
+  def self.ransackable_attributes(_)
+    ["created_at", "id", "name", "updated_at"]
+  end
 
   def full_name
     out = ""
