@@ -1,5 +1,5 @@
 class SchoolNameDuplicate < ApplicationRecord
-  validates_uniqueness_of :name
+  validates_uniqueness_of :name, case_sensitive: false
 
   strip_attributes
 
