@@ -1,6 +1,6 @@
 class TrackableEvent < ApplicationRecord
   validates_presence_of :band_id, :trackable_tag, :user
-  validates_uniqueness_of :band_id, scope: :trackable_tag_id, if: -> { !trackable_tag&.allow_duplicate_band_events }
+  validates_uniqueness_of :band_id, case_sensitive: false, scope: :trackable_tag_id, if: -> { !trackable_tag&.allow_duplicate_band_events }
 
   strip_attributes
 
