@@ -22,8 +22,7 @@ class OauthTest < ActionDispatch::IntegrationTest
 
     should "return data for questionnaire" do
       get questionnaires_path, headers: auth_headers(@token.token)
-      assert_response :redirect
-      assert_redirected_to new_user_session_url
+      assert_response :unauthorized
     end
   end
 
