@@ -3,7 +3,7 @@ require 'test_helper'
 class TrackableTagTest < ActiveSupport::TestCase
   should strip_attribute :name
   should validate_presence_of :name
-  should validate_uniqueness_of :name
+  should validate_uniqueness_of(:name).ignoring_case_sensitivity
   should have_many :trackable_events
 
   should "implement sorted_events" do
